Fix playback tracking when switching media via the queue
Also, improve reliability of tracking by extending Babbling's
PlaybackTracker. Previously, if you issued a "stop" the current
position/media might not actually get logged, so that's patched.

As part of this change, we switch to using contentId as the actual media
ID and contentUrl for the playback URL. This is more canonical anyway,
and makes it simpler to associate chromecast state with our local media
representations.