Fix default media player queue support by keeping the server alive
Per code comments, when streaming via ranges, if we get to the end
of a video we will likely fetch the last range and then not need to
fetch anything for some time, so the server would, by default, see
that as the client going away and stop serving.

By allowing Playables to have "clients" attached to them to keep them
alive, we can ensure the Server will stick around in case the user
wants to continue through the queue.