Add explicit option to allow Shougun to keep the process alive
Shougun components can use this for optimization purposes—in particular,
the LocalDiscovery's changes() proc can make use of fsevents for
watching file changes if *and only if* we're allowed to keep the process
alive. If not, we *must* use polling in order to still get events while
not keeping the process alive.