Support selecting a preferred audio language (#20)
* Experiment with support for a "preferred language" pref for Media

Ideally we could just let the player handle this, but for at least one
media file I have in MKV format, the Chromecast app reports that there
are no audio tracks at all, so... sadly, for widest compat, we have to
fallback to ffmpeg. Future work could explore the Shougun Player-based
solution for other container types....

Still TODO:

- Probably, override any stored preference in the PlaybackOptions
- Store and retrieve preferences
- CLI support?

* Support providing prefs with IPlaybackOptions

* Support receiving IPlaybackOptions over RPC

* Add methods of reading and writing prefs from and to the DB

* Fix sqlite3 schema version

* Inflate prefs on play()

We *could* do it on query, but this is way simpler and way more
efficient. Right now there's not really a reason to eagerly fetch
preferences, so let's not bother.

* Expose IPrefsTracker on the Shougun object

Also, have IPrefsTracker return the updated prefs

* Unify Prefs interfaces; support deleting prefs