Implement RPC v2 with provider errors, cursor support (#43)
* Begin initial implemetnation of a custom msgpack-rpc server

* Ensure msgIDs wrap properly if necessary

* Switch to using our custom msgpack rpc server

* Add some extra safety to the msgpack rpc server

* Prepare a Version-negotiation RPC handler

* Fix VersionNegotiatorFactory-test

* Add some simple tests for the VersionNegotiatorFactory

* Support "composing" RPC handlers with support for changed signatures

The whole point of versioned RPC is that we can change the signature of
a method in a backwards-incompatible way in new versions. It's a bit
tricky with typescript, but doable...

* Implement queryRecent/queryRecommended v2, returning provider errors

Also supports cursors!

* Fix error routing

I should have known that the previous approach was too easy. Ah well, at
least this way is slightly more consistent so probably for the best

* Implement search v2

* Clean up some lint errors

* Clean up remaining lint errors

* Provide compatibility layer with broken v1 clients

As mentioned in the comments, the shougun/cli client currently wraps
params in an extra array. None of the v1 methods expect any array
arguments, so we can safely unpack this if the client is targetting API
v1, while keeping things clean for new v2 clients.