Attempt to improve support for fast-forwarding in queued media (#47)
* Clean up unnecessary condition

* Clean up error string format

* Attempt to load duration for for queue media

I think this will help with skipping in non-natively-playable media
types. It seems like I'm mostly encountering the issue with being unable
to skip properly after going to the "next" episode; currently, only the
explicitly requested media has the duration, so *perhaps* the Chromecast
isn't passing the proper timestamp if it's past whatever it believes is
the duration of the video (which is basically however long it's played).