Attempt to "inflate" media directly for RPC playback
Our previous attempt repeated the search locally with the given title to
get the search result object, since that's how we interop with external
playables. However, there's non-trivial latency required there, and it
also doesn't always work—in particular for Last Week Tonight.

This approach leans on babbling's `playUrl` functionality, which we have
in the media's ID, to attempt to play the requested media directly and
skip the search step. In theory this should always work, but we still
fallback to the search-based approach *just in case*.

At some point we should consider separating the `IMedia` interface to
more clearly describe when it's a serialized media object from eg RPC vs
an *actual* object returned by a Queryable that can be safely played
directly....