Attempt to "inflate" media directly for RPC playback (#48)
Our previous attempt repeated the search locally with the given title to
get the search result object, since that object is how we interop with external
playables. However, there's non-trivial latency involved in performing this search—usually about a second, but potentially longer since we're querying all connected services. Of course, the latency involved in starting the Chromecast app and starting playback completely dwarfs that (about 20s for HBO, for example) but a win's a win.

This approach leans on babbling's `playUrl` functionality, which we have
in the media's ID, to attempt to play the requested media directly and
skip the search step. In theory this should always work, but we still
fallback to the search-based approach *just in case*.

At some point we should consider separating the `IMedia` interface to
more clearly describe when it's a serialized media object from eg RPC vs
an *actual* object returned by a Queryable that can be safely played
directly....